[DDW-169] disable cross-compilation between main and renderer process
