+++ source/main/ipc-api/compress-logs.js
@@ -5,13 +5,7 @@ import path from 'path';
 import { get } from 'lodash';
 import { appLogsFolderPath, pubLogsFolderPath } from '../config';
 import { Logger, stringifyError } from '../../renderer/app/utils/logging';
-
-const CHANNEL_NAME = 'compress-logs';
-
-export const COMPRESS_LOGS = {
-  REQUEST: CHANNEL_NAME,
-  SUCCESS: `${CHANNEL_NAME}-success`,
-};
+import { COMPRESS_LOGS } from '../../common/ipc-api';
 
 export default () => {
   ipcMain.on(COMPRESS_LOGS.REQUEST, (event, logs) => {
