+++ source/renderer/app/stores/ProfileStore.js
@@ -9,10 +9,7 @@ import Request from './lib/LocalizedRequest';
 import environment from '../../../common/environment';
 import { THEMES } from '../themes/index';
 import { ROUTES } from '../routes-config';
-import { GET_LOGS } from '../../../main/ipc-api/get-logs';
-import { DOWNLOAD_LOGS } from '../../../main/ipc-api/download-logs';
-import { COMPRESS_LOGS } from '../../../main/ipc-api/compress-logs';
-import { DELETE_COMPRESSED_LOGS } from '../../../main/ipc-api/delete-compressed-logs';
+import { GET_LOGS, DOWNLOAD_LOGS, COMPRESS_LOGS, DELETE_COMPRESSED_LOGS } from '../../../common/ipc-api';
 import LocalizableError from '../i18n/LocalizableError';
 import globalMessages from '../i18n/global-messages';
 import { WalletSupportRequestLogsCompressError } from '../i18n/errors';
