+++ source/renderer/index.js
@@ -1,1 +1,2 @@
 import './app/index';
+import '../main/index';
